refactor(exception): exception 하위 패키지 추가 및 재분배
domain.global.exception 패키지에 다음 하위 패키지를 추가했습니다:
- error: response 관련 클래스 집합
- custom: Http 상태 코드 별 Exception 구현체 집합
