--- src/main/java/roomescape/domain/global/exception/BaseException.java
@@ -1,6 +1,0 @@
-package roomescape.domain.global.exception;
-
-public interface BaseException {
-
-    ErrorCode getErrorCode();
-}
