+++ src/main/java/roomescape/domain/global/exception/custom/BaseException.java
@@ -1,0 +1,8 @@
+package roomescape.domain.global.exception.custom;
+
+import roomescape.domain.global.exception.error.ErrorCode;
+
+public interface BaseException {
+
+    ErrorCode getErrorCode();
+}
