+++ src/main/java/roomescape/domain/global/exception/custom/NotFoundException.java
@@ -1,4 +1,6 @@
-package roomescape.domain.global.exception;
+package roomescape.domain.global.exception.custom;
+
+import roomescape.domain.global.exception.error.ErrorCode;
 
 public class NotFoundException extends RuntimeException implements BaseException {
