+++ src/main/java/roomescape/domain/global/exception/error/ErrorDetail.java
@@ -1,4 +1,4 @@
-package roomescape.domain.global.exception;
+package roomescape.domain.global.exception.error;
 
 public record ErrorDetail(String field, String value, String message) {
