+++ src/main/java/roomescape/domain/reservation/service/ReservationService.java
@@ -5,13 +5,13 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
-import roomescape.domain.global.exception.BadRequestException;
-import roomescape.domain.global.exception.ConflictException;
-import roomescape.domain.global.exception.ErrorCode;
-import roomescape.domain.global.exception.ErrorDetail;
-import roomescape.domain.global.exception.ForbiddenException;
-import roomescape.domain.global.exception.NotFoundException;
-import roomescape.domain.global.exception.UnprocessableEntityException;
+import roomescape.domain.global.exception.custom.BadRequestException;
+import roomescape.domain.global.exception.custom.ConflictException;
+import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.domain.global.exception.error.ErrorDetail;
+import roomescape.domain.global.exception.custom.ForbiddenException;
+import roomescape.domain.global.exception.custom.NotFoundException;
+import roomescape.domain.global.exception.custom.UnprocessableEntityException;
 import roomescape.domain.reservation.dto.request.ReservationCreateRequestDto;
 import roomescape.domain.reservation.dto.request.ReservationUpdateRequestDto;
 import roomescape.domain.reservation.dto.response.ReservationCreateResponseDto;
