+++ src/main/java/roomescape/domain/reservation/validator/ReservationValidator.java
@@ -2,9 +2,9 @@ package roomescape.domain.reservation.validator;
 
 import java.util.ArrayList;
 import java.util.List;
-import roomescape.domain.global.exception.BadRequestException;
-import roomescape.domain.global.exception.ErrorCode;
-import roomescape.domain.global.exception.ErrorDetail;
+import roomescape.domain.global.exception.custom.BadRequestException;
+import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.domain.global.exception.error.ErrorDetail;
 
 public final class ReservationValidator {
