+++ src/main/java/roomescape/domain/theme/service/ThemeService.java
@@ -3,10 +3,10 @@ package roomescape.domain.theme.service;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.global.exception.ConflictException;
-import roomescape.domain.global.exception.ErrorCode;
-import roomescape.domain.global.exception.NotFoundException;
-import roomescape.domain.global.exception.UnprocessableEntityException;
+import roomescape.domain.global.exception.custom.ConflictException;
+import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.domain.global.exception.custom.NotFoundException;
+import roomescape.domain.global.exception.custom.UnprocessableEntityException;
 import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.theme.dto.request.ThemeCreateRequestDto;
 import roomescape.domain.theme.dto.response.ThemeResponseDto;
