+++ src/main/java/roomescape/domain/theme/validator/ThemeValidator.java
@@ -2,9 +2,9 @@ package roomescape.domain.theme.validator;
 
 import java.util.ArrayList;
 import java.util.List;
-import roomescape.domain.global.exception.BadRequestException;
-import roomescape.domain.global.exception.ErrorCode;
-import roomescape.domain.global.exception.ErrorDetail;
+import roomescape.domain.global.exception.custom.BadRequestException;
+import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.domain.global.exception.error.ErrorDetail;
 
 public final class ThemeValidator {
