+++ src/main/java/roomescape/domain/time/service/TimeService.java
@@ -4,12 +4,12 @@ import java.time.Clock;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.domain.global.exception.BadRequestException;
-import roomescape.domain.global.exception.ConflictException;
-import roomescape.domain.global.exception.ErrorCode;
-import roomescape.domain.global.exception.ErrorDetail;
-import roomescape.domain.global.exception.NotFoundException;
-import roomescape.domain.global.exception.UnprocessableEntityException;
+import roomescape.domain.global.exception.custom.BadRequestException;
+import roomescape.domain.global.exception.custom.ConflictException;
+import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.domain.global.exception.error.ErrorDetail;
+import roomescape.domain.global.exception.custom.NotFoundException;
+import roomescape.domain.global.exception.custom.UnprocessableEntityException;
 import roomescape.domain.reservation.repository.ReservationRepository;
 import roomescape.domain.theme.repository.ThemeRepository;
 import roomescape.domain.time.dto.request.TimeCreateRequestDto;
