+++ src/test/java/roomescape/domain/reservation/dto/request/ReservationCreateRequestDtoTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.global.exception.BadRequestException;
+import roomescape.domain.global.exception.custom.BadRequestException;
 
 class ReservationCreateRequestDtoTest {
