+++ src/test/java/roomescape/domain/reservation/entity/ReservationTest.java
@@ -11,8 +11,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.global.exception.ErrorCode;
-import roomescape.domain.global.exception.UnprocessableEntityException;
+import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.domain.global.exception.custom.UnprocessableEntityException;
 import roomescape.domain.theme.entity.Theme;
 import roomescape.domain.time.entity.Time;
 import roomescape.global.ExceptionAssertions;
