+++ src/test/java/roomescape/domain/theme/dto/request/ThemeCreateRequestDtoTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.global.exception.BadRequestException;
-import roomescape.domain.global.exception.ErrorCode;
+import roomescape.domain.global.exception.custom.BadRequestException;
+import roomescape.domain.global.exception.error.ErrorCode;
 import roomescape.global.ExceptionAssertions;
 
 class ThemeCreateRequestDtoTest {
