+++ src/test/java/roomescape/domain/theme/service/ThemeServiceTest.java
@@ -16,9 +16,9 @@ import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.global.exception.ConflictException;
-import roomescape.domain.global.exception.ErrorCode;
-import roomescape.domain.global.exception.NotFoundException;
+import roomescape.domain.global.exception.custom.ConflictException;
+import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.domain.global.exception.custom.NotFoundException;
 import roomescape.domain.reservation.entity.Reservation;
 import roomescape.domain.reservation.repository.FakeReservationRepository;
 import roomescape.domain.reservation.repository.ReservationRepository;
