+++ src/test/java/roomescape/domain/theme/validator/ThemeValidatorTest.java
@@ -6,8 +6,8 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.global.exception.ErrorCode;
-import roomescape.domain.global.exception.ErrorDetail;
+import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.domain.global.exception.error.ErrorDetail;
 import roomescape.global.ExceptionAssertions;
 
 class ThemeValidatorTest {
