+++ src/test/java/roomescape/domain/time/service/TimeServiceTest.java
@@ -13,10 +13,10 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import roomescape.domain.global.exception.BadRequestException;
-import roomescape.domain.global.exception.ConflictException;
-import roomescape.domain.global.exception.ErrorCode;
-import roomescape.domain.global.exception.UnprocessableEntityException;
+import roomescape.domain.global.exception.custom.BadRequestException;
+import roomescape.domain.global.exception.custom.ConflictException;
+import roomescape.domain.global.exception.error.ErrorCode;
+import roomescape.domain.global.exception.custom.UnprocessableEntityException;
 import roomescape.domain.reservation.entity.Reservation;
 import roomescape.domain.reservation.repository.FakeReservationRepository;
 import roomescape.domain.reservation.repository.ReservationRepository;
